AuraPy/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
comtypes==1.4.6
idna==3.8
MouseInfo==0.1.3
newsapi-python==0.2.7
PyAudio==0.2.14
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pyttsx3==2.91
pytweening==1.2.0
pywin32==306
requests==2.32.3
setuptools==74.1.2
SpeechRecognition==3.10.4
typing_extensions==4.12.2
urllib3==2.2.2
